Should complete #318
